refactor: reorganize backend into layered architecture
