+++ backend/tests/routes/config.test.js
@@ -4,7 +4,7 @@ const request = require('supertest');
 process.env.JWT_SECRET = 'a'.repeat(32);
 process.env.DATABASE_URL = 'postgresql://user:pass@localhost:5432/test';
 process.env.GOOGLE_CLIENT_ID = 'dummy';
-const { app } = require('../../src/app');
+const { app } = require('../../src/presentation/app');
 
 test('GET /api/config/google returns client ID when configured', async () => {
   const originalClientId = process.env.GOOGLE_CLIENT_ID;
